refactor(navbar): move theme logic out of component to a hook
